feature #907 sprint1 :   Update the SQL scripts that create the tables for the application Suggestion Box: int becomes an int8 (or bigint or number(19,0))

+++ suggestion-box/suggestion-box-config/src/main/config/dbRepository/oracle/suggestion-box/001/create_table.sql
@@ -1,19 +1,19 @@
 CREATE TABLE sc_suggestion_box (
-  id              VARCHAR(40) PRIMARY KEY,
-  instanceId      VARCHAR(30) NOT NULL,
-  createDate      TIMESTAMP   NOT NULL,
-  createdBy       VARCHAR(40) NOT NULL,
+  id              VARCHAR(40)  PRIMARY KEY,
+  instanceId      VARCHAR(30)  NOT NULL,
+  createDate      TIMESTAMP    NOT NULL,
+  createdBy       VARCHAR(40)  NOT NULL,
   lastUpdateDate  TIMESTAMP,
   lastUpdatedBy   VARCHAR(40),
-  version         int        NOT NULL
+  version         NUMBER(19,0) NOT NULL
 );
 
 CREATE TABLE sc_suggestion (
-  id              VARCHAR(40) PRIMARY KEY,
-  suggestionBoxId VARCHAR(40) NOT NULL,
-  createDate      TIMESTAMP   NOT NULL,
-  createdBy       VARCHAR(40) NOT NULL,
+  id              VARCHAR(40)  PRIMARY KEY,
+  suggestionBoxId VARCHAR(40)  NOT NULL,
+  createDate      TIMESTAMP    NOT NULL,
+  createdBy       VARCHAR(40)  NOT NULL,
   lastUpdateDate  TIMESTAMP,
   lastUpdatedBy   VARCHAR(40),
-  version         int        NOT NULL
+  version         NUMBER(19,0) NOT NULL
 );
